+++ src/ngraph_api.cc
@@ -33,13 +33,13 @@ bool ngraph_is_enabled() { return IsEnabled(); }
 
 size_t ngraph_backends_len() { return BackendsLen(); }
 
-bool ngraph_list_backends(char** backends, int backends_len) {
+bool ngraph_list_backends(char** backends, size_t backends_len) {
   const auto ngraph_backends = ListBackends();
-  if ((unsigned)backends_len != ngraph_backends.size()) {
+  if (backends_len != ngraph_backends.size()) {
     return false;
   }
 
-  for (size_t idx = 0; idx < (unsigned)backends_len; idx++) {
+  for (size_t idx = 0; idx < backends_len; idx++) {
     backends[idx] = strdup(ngraph_backends[idx].c_str());
   }
   return true;
